\name{CFC}
\alias{CFC}
\docType{data}
\title{
Annual mean of cloud fraction cover
}
\description{
This dataset contains a raster of annual mean of cloud fraction cover
}
\usage{
require(raster)
data(CFC)
}
\format{ }
\details{
Annual mean of cloud fraction cover is used as a proxy of a number of potential clear-sky days at given location.
}
\source{

}
\references{

}
\examples{
require(raster)
data(CFC)
str(CFC)

}
\keyword{datasets}
