\name{cstRead}
\alias{cstRead}

\title{
Read values of clear sky transmissivity
}
\description{
Read values of clear sky transmissivity map for a given locations (in lat/lon)
}
\usage{
cstRead(lat,lon)
}

\arguments{
  \item{lat}{
Latitude in decimal degrees.
}
  \item{lon}{
Longitude in decimal degrees.
}
}

%\details{}

\value{
Clear sky transmissivity 
}

%\references{}

\author{
Jedrzej S. Bojanowski
}

%\note{}

\seealso{
'cst'
}
\examples{
cstRead(50,16)
}

%\keyword{}

