\name{dayOfYear}
\alias{dayOfYear}

\title{
Convert 'Date' to number of day in a year 
}
\description{
Function gives a day number of the year (julian day of the year) based on the date in class 'Date'.
}
\usage{
dayOfYear(dat)
}

\arguments{
  \item{dat}{
Date in class 'Date'.
}
}

%\details{}

\value{
Numeric number of day in a year. }

%\references{}

\author{
Jedrzej S. Bojanowski
}
%\note{}

%\seealso{}

\examples{

dayOfYear(as.Date("2009-01-11"))

}

%\keyword{}

