\name{degrees}
\alias{degrees}

\title{
Convert radians to degrees
}
\description{
Converts radians to degrees
}
\usage{
degrees(radians)
}

\arguments{
\item{radians}{numeric}

}

%\details{}

\value{
Degrees.
}

%\references{}

\author{
Jedrzej S. Bojanowski
}

%\note{}

\seealso{
'radians'
}
\examples{
degrees(0.95)
}

%\keyword{}

