\name{radians}
\alias{radians}

\title{
Convert degrees to radians
}
\description{
Converts degrees to radians
}
\usage{
radians(degrees)
}

\arguments{
\item{degrees}{numeric}
}

%\details{}

\value{
Radians.
}

%\references{}

\author{
Jedrzej S. Bojanowski
}

%\note{}

\seealso{
'degrees'
}
\examples{
radians(55)
}

%\keyword{}

